Add files via upload
Added a function that when inject GorillaInternal is pressed, it checks for common path's that hl2.exe locates in, then it runs it, but if it's not found, it will get you to a window where you can type in the path to your hl2.exe, and then inject the GorillaInternal to there, also you can choose if you want to inject to an already open process of hl2.exe(Team Fortress 2). Also added checks if the GorillaInternal.dll exists in the path that the GorillaMaintainer is in, if it doesn't exist,  it will say it in the information text field, and it will not open TF2, nor attempt to inject the DLL. Then you can download the GorillaInternal.dll using the "Download GorillaInternal.dll" button.