Add files via upload
Added a page for  selecting launch options for Team Fortress 2, and saving and loading the options to/from a JSON file, auto loading the settings on the start of the program, added buttons for opening, and closing Team Fortress 2, added a checkBox to GorillaInjector page for choosing whether to inject to Team Fortress 2, or to the exe you will choose from the fileExplorer, added this since it's frustrating to always choose the hl2.exe from the fileExplorer. Made it like the program waits until Team Fortress 2 has loaded, and then inject(might not work the best, lets see).